// core/src/experimental/input/slider/main/slider.api.ts
/*
 * SPDX-License-Identifier: EUPL-1.2 OR LicenseRef-commercial
 *
 * Copyright (c) 2012-2026 mgm technology partners GmbH
 *
 * Dual License
 * ------------
 * This source file is part of the mgm A12 Platform and available under
 * a choice of two different licenses:
 *
 * 1. Open-Source License - EUPL v1.2
 *    You may redistribute and/or modify this file under the terms of the
 *    European Union Public License, version 1.2 - see https://eupl.eu/.
 *
 * 2. Commercial License
 *    Alternatively, you may obtain a commercial license from
 *    mgm technology partners GmbH, that permits use of this software
 *    under different terms (including support and maintenance services).
 *
 *    Please contact a12-license@mgm-tp.com for more information.
 *
 * You must select and comply with exactly one of the above license options.
 *
 * Warranty Disclaimer (applies to either option)
 * ----------------------------------------------
 * THIS SOFTWARE IS PROVIDED "AS IS" AND WITHOUT WARRANTY OF ANY KIND,
 * WHETHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
 * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
 * NON-INFRINGEMENT, EXCEPT WHERE SUCH DISCLAIMERS ARE HELD TO BE
 * LEGALLY INVALID. SEE THE RESPECTIVE LICENSE TEXT FOR DETAILS.
 */

import type { BaseInputProps } from "../../../../input/base/template/base.tpl.api.js";

declare type OmittedBaseInputProps =
	| "warning"
	| "warningMessage"
	| "helperText"
	| "tooltips"
	| "ariaDescribedby"
	| "breakTooltipsToNewLine";

export interface SliderProps extends Omit<BaseInputProps, OmittedBaseInputProps> {
	/**
	 * An array that contains all marks of the slider.
	 */
	marks: SliderMark[];

	/**
	 * The currently selected value.
	 */
	value?: string;

	/**
	 * A label to show on the left of the slider, normally used for indicating left-most value.
	 */
	leftLabel?: string;

	/**
	 * A label to show on the right of the slider, normally used for indicating right-most value.
	 */
	rightLabel?: string;

	/**
	 * A handler function that will be fired when the mouse is released.
	 * @param value – {@link SliderMark}'s value
	 */
	onChange?(value: string): void;
}

export interface SliderMark {
	/**
	 * The label of the slider mark.
	 */
	label?: string;

	/**
	 * Corresponding value of the mark that will be returned when onChange is fired.
	 */
	value: string;

	/**
	 * Specifies if the mark is disabled.
	 */
	disabled?: boolean;
}